*ActorHit a pointer to a empty AActor class that gets pouplated on hit
made an if statement to say if Actorhit which mean if ActorHit is true aka is no false which means its not null then do this statement, which gets the name from ActorHit and logs it out.
this had to be done otherwise UE4 would crash as it was accesing a null pointer.